Move rest of Valkyrie-specific code to lib.
