+++ app/controllers/collections_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 class CollectionsController < ApplicationController
-  include ModelControllerBehavior
+  include Valkyrie::ControllerConcerns::ModelControllerBehavior
   self.form_class = CollectionForm
   self.resource_class = Collection
