+++ app/controllers/pages_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 class PagesController < ApplicationController
-  include ModelControllerBehavior
+  include Valkyrie::ControllerConcerns::ModelControllerBehavior
   self.form_class = PageForm
   self.resource_class = Page
