+++ config/initializers/register_adapters.rb
@@ -17,7 +17,7 @@ Rails.application.config.to_prepare do
 
   Valkyrie::Adapter.register(
     Valkyrie::Persistence::Solr::Adapter.new(connection: Blacklight.default_index.connection,
-                                             resource_indexer: Valkyrie::AccessControlsIndexer),
+                                             resource_indexer: Valkyrie::Indexers::AccessControlsIndexer),
     :index_solr
   )
 
@@ -31,23 +31,23 @@ Rails.application.config.to_prepare do
     :memory
   )
 
-  persister_list = Valkyrie::DecoratorList.new(
-    Valkyrie::DecoratorWithArguments.new(FileSetAppendingPersister,
-                                         repository: Valkyrie.config.storage_adapter,
-                                         node_factory: FileNode,
-                                         file_container_factory: FileSet),
+  persister_list = Valkyrie::Decorators::DecoratorList.new(
+    Valkyrie::Decorators::DecoratorWithArguments.new(FileSetAppendingPersister,
+                                                     repository: Valkyrie.config.storage_adapter,
+                                                     node_factory: FileNode,
+                                                     file_container_factory: FileSet),
     ParentCleanupPersister,
     AppendingPersister
   )
 
   Valkyrie::Adapter.register(
-    AdapterContainer.new(persister: persister_list.new(
+    Valkyrie::AdapterContainer.new(persister: persister_list.new(
       CompositePersister.new(
         Valkyrie.config.adapter.persister,
         Valkyrie::Adapter.find(:index_solr).persister
       )
     ),
-                         query_service: Valkyrie.config.adapter.query_service),
+                                   query_service: Valkyrie.config.adapter.query_service),
     :indexing_persister
   )
 end
