+++ lib/valkyrie/adapter.rb
