+++ lib/valkyrie/file_repository.rb
