+++ lib/valkyrie/id.rb
