+++ lib/valkyrie/model/access_controls.rb
