+++ lib/valkyrie/persistence/active_fedora/dynamic_klass.rb
