+++ lib/valkyrie/persistence/active_fedora/orm/resource.rb
