+++ lib/valkyrie/persistence/active_fedora/queries/default_paginator.rb
