+++ lib/valkyrie/persistence/active_fedora/queries/find_all_query.rb
