+++ lib/valkyrie/persistence/active_fedora/query_service.rb
