+++ lib/valkyrie/persistence/active_fedora/resource_factory.rb
