--- lib/valkyrie/persistence/fedora/queries/default_paginator.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-module Valkyrie::Persistence::Fedora::Queries
-  class DefaultPaginator
-    def next_page
-      1
-    end
-
-    def per_page
-      100
-    end
-
-    def has_next?
-      true
-    end
-  end
-end
