+++ lib/valkyrie/types.rb
