+++ spec/valkyrie/persistence/active_fedora/adapter_spec.rb
@@ -2,7 +2,7 @@
 require 'rails_helper'
 require 'valkyrie/specs/shared_specs'
 
-RSpec.describe Valkyrie::Persistence::Fedora::Adapter do
+RSpec.describe Valkyrie::Persistence::ActiveFedora::Adapter do
   let(:adapter) { described_class }
   it_behaves_like "a Valkyrie::Adapter"
 end
