+++ spec/valkyrie/persistence/active_fedora/persister_spec.rb
@@ -2,7 +2,7 @@
 require 'rails_helper'
 require 'valkyrie/specs/shared_specs'
 
-RSpec.describe Valkyrie::Persistence::Fedora::Persister do
+RSpec.describe Valkyrie::Persistence::ActiveFedora::Persister do
   let(:persister) { described_class }
   it_behaves_like "a Valkyrie::Persister"
 end
